# modules/api_contract/openapi.yaml
openapi: 3.0.0
info:
  title: Azure Tenant Grapher Remote API
  version: 1.0.0
  description: |
    REST API for remote execution of Azure Tenant Grapher operations.
    Supports long-running async operations with job polling and progress updates.

    **Key Features:**
    - Async job queue for long-running operations (20+ min scans)
    - Real-time progress updates via SSE
    - API key authentication
    - All CLI commands accessible via REST

    **Authentication:**
    All endpoints require `X-API-Key` header with valid API key.

    **Job Lifecycle:**
    1. Submit job (POST /jobs/{operation})
    2. Poll status (GET /jobs/{job_id})
    3. Get results (GET /jobs/{job_id}/results)
    4. Stream progress (GET /jobs/{job_id}/progress - SSE)
servers:
  - url: https://api.atg.example.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - ApiKeyAuth: []
paths:
...
  /jobs/scan:
    post:
      summary: Submit scan job
      operationId: submit_scan_job
      tags:
        - Jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanJobRequest'
      responses:
        '202':
          $ref: '#/components/responses/JobAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
...
  schemas:
...
    ScanJobRequest:
      type: object
      required:
        - tenant_id
      properties:
        tenant_id:
          type: string
          description: Azure tenant ID
        resource_limit:
          type: integer
          nullable: true
          description: Maximum number of resources to process
        max_llm_threads:
          type: integer
          default: 5
        max_build_threads:
          type: integer
          default: 20
        max_retries:
          type: integer
          default: 3
        max_concurrency:
          type: integer
          default: 100
        generate_spec:
          type: boolean
          default: false
          description: Generate tenant spec after scan
        visualize:
          type: boolean
          default: false
          description: Generate visualization after scan
        rebuild_edges:
          type: boolean
          default: false
        no_aad_import:
          type: boolean
          default: false
        filter_by_subscriptions:
          type: string
          nullable: true
          description: Comma-separated subscription IDs
        filter_by_rgs:
          type: string
          nullable: true
          description: Comma-separated resource group names
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code (e.g., INVALID_TENANT_ID, JOB_NOT_FOUND)
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              additionalProperties: true
              description: Additional error details
  responses:
    JobAccepted:
      description: Job accepted and queued
      content:
        application/json:
          schema:
            type: object
            required:
              - job_id
              - status
            properties:
              job_id:
                type: string
                format: uuid
              status:
                type: string
                enum: [pending]
              status_url:
                type: string
                format: uri
                description: URL to poll job status
              progress_url:
                type: string
                format: uri
                description: SSE endpoint for progress streaming
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...